# backend/app/models/domain.py
from enum import Enum as PyEnum


class ProblemDomain(str, PyEnum):
    WATER_MANAGEMENT = "WATER_MANAGEMENT"
    ROADS_INFRASTRUCTURE = "ROADS_INFRASTRUCTURE"
    HEALTHCARE_SANITATION = "HEALTHCARE_SANITATION"
    EDUCATION = "EDUCATION"
    AGRICULTURE = "AGRICULTURE"
    ENERGY_ELECTRICITY = "ENERGY_ELECTRICITY"
    ENVIRONMENT_WASTE = "ENVIRONMENT_WASTE"
    PUBLIC_SAFETY = "PUBLIC_SAFETY"
    GOVERNANCE_SERVICES = "GOVERNANCE_SERVICES"
    OTHER = "OTHER"