# docs/openapi-demo-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Samanvay API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/problems": {
      "get": {
        "tags": [
          "Student Problem Feed"
        ],
        "summary": "List standardized societal problems",
        "description": "Student discovery feed for standardized problems. Supports filtering by domain, geographic radius (Haversine), minimum priority score, and status. Zero citizen PII is exposed.",
        "operationId": "get_standardized_problems_problems_get",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProblemDomain"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by citizen-selected domain",
              "title": "Domain"
            },
            "description": "Filter by citizen-selected domain"
          },
          {
            "name": "latitude",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90.0,
                  "minimum": -90.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Center latitude for distance filtering",
              "title": "Latitude"
            },
            "description": "Center latitude for distance filtering"
          },
          {
            "name": "longitude",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180.0,
                  "minimum": -180.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Center longitude for distance filtering",
              "title": "Longitude"
            },
            "description": "Center longitude for distance filtering"
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100.0,
                  "minimum": 1.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search radius in kilometers (default: 25 km, max: 100 km)",
              "title": "Radius Km"
            },
            "description": "Search radius in kilometers (default: 25 km, max: 100 km)"
          },
          {
            "name": "min_priority_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100.0,
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum priority score (0-100)",
              "title": "Min Priority Score"
            },
            "description": "Minimum priority score (0-100)"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProblemStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Problem status (defaults to OPEN & UNDER_INVESTIGATION)",
              "title": "Status"
            },
            "description": "Problem status (defaults to OPEN & UNDER_INVESTIGATION)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Pagination limit (max: 50)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Pagination limit (max: 50)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/solutions": {
      "post": {
        "tags": [
          "Solutions & Industry Review"
        ],
        "summary": "Propose a solution for an adopted problem",
        "description": "Allows authenticated and verified UNIVERSITY mentors to propose a technical solution for their team.",
        "operationId": "create_solution_solutions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolutionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolutionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/solutions/{solution_id}/industry-review": {
      "post": {
        "tags": [
          "Solutions & Industry Review"
        ],
        "summary": "Review a proposed solution",
        "description": "Allows verified INDUSTRY users to approve, request changes, or reject a submitted solution.",
        "operationId": "review_solution_solutions__solution_id__industry_review_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "solution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Solution Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndustryReviewCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndustryReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/problems": {
      "get": {
        "tags": [
          "Public APIs & Analytics"
        ],
        "summary": "Public Societal Problems",
        "description": "Public feed of standardized societal problems and status. Zero citizen PII is exposed.",
        "operationId": "get_public_problems_public_problems_get",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProblemDomain"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by problem domain",
              "title": "Domain"
            },
            "description": "Filter by problem domain"
          },
          {
            "name": "latitude",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90.0,
                  "minimum": -90.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Center latitude for distance filter",
              "title": "Latitude"
            },
            "description": "Center latitude for distance filter"
          },
          {
            "name": "longitude",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180.0,
                  "minimum": -180.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Center longitude for distance filter",
              "title": "Longitude"
            },
            "description": "Center longitude for distance filter"
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100.0,
                  "minimum": 1.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Radius in km (default 25, max 100)",
              "title": "Radius Km"
            },
            "description": "Radius in km (default 25, max 100)"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProblemStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Problem status filter",
              "title": "Status"
            },
            "description": "Problem status filter"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Pagination limit (max: 50)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Pagination limit (max: 50)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndustryReviewCreate": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/IndustryReviewDecision"
          },
          "review_comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Comment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "IndustryReviewCreate"
      },
      "IndustryReviewDecision": {
        "type": "string",
        "enum": [
          "APPROVE",
          "NEEDS_CHANGES",
          "REJECT"
        ],
        "title": "IndustryReviewDecision"
      },
      "IndustryReviewResponse": {
        "properties": {
          "solution_id": {
            "type": "string",
            "format": "uuid",
            "title": "Solution Id"
          },
          "status": {
            "$ref": "#/components/schemas/SolutionStatus"
          },
          "industry_review_status": {
            "$ref": "#/components/schemas/IndustryReviewStatus"
          },
          "decision": {
            "$ref": "#/components/schemas/IndustryReviewDecision"
          }
        },
        "type": "object",
        "required": [
          "solution_id",
          "status",
          "industry_review_status",
          "decision"
        ],
        "title": "IndustryReviewResponse"
      },
      "IndustryReviewStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED"
        ],
        "title": "IndustryReviewStatus"
      },
      "ProblemDomain": {
        "type": "string",
        "enum": [
          "WATER_MANAGEMENT",
          "ROADS_INFRASTRUCTURE",
          "HEALTHCARE_SANITATION",
          "EDUCATION",
          "AGRICULTURE",
          "ENERGY_ELECTRICITY",
          "ENVIRONMENT_WASTE",
          "PUBLIC_SAFETY",
          "GOVERNANCE_SERVICES",
          "OTHER"
        ],
        "title": "ProblemDomain"
      },
      "ProblemListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProblemRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ProblemListResponse"
      },
      "ProblemRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "problem_summary": {
            "type": "string",
            "title": "Problem Summary"
          },
          "affected_community": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Community"
          },
          "observed_impact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Impact"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score"
          },
          "report_count": {
            "type": "integer",
            "title": "Report Count"
          },
          "evidence_count": {
            "type": "integer",
            "title": "Evidence Count"
          },
          "status": {
            "$ref": "#/components/schemas/ProblemStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "domain",
          "problem_summary",
          "priority_score",
          "report_count",
          "evidence_count",
          "status",
          "created_at"
        ],
        "title": "ProblemRead"
      },
      "ProblemStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "UNDER_INVESTIGATION",
          "ADOPTED",
          "IN_DEVELOPMENT",
          "PROTOTYPE",
          "GOVERNMENT_REVIEW",
          "RESOLVED"
        ],
        "title": "ProblemStatus"
      },
...
      "PublicProblemListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicProblemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PublicProblemListResponse"
      },
      "PublicProblemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "problem_summary": {
            "type": "string",
            "title": "Problem Summary"
          },
          "observed_impact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Impact"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score"
          },
          "report_count": {
            "type": "integer",
            "title": "Report Count"
          },
          "evidence_count": {
            "type": "integer",
            "title": "Evidence Count"
          },
          "status": {
            "$ref": "#/components/schemas/ProblemStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "domain",
          "problem_summary",
          "priority_score",
          "report_count",
          "evidence_count",
          "status",
          "created_at"
        ],
        "title": "PublicProblemResponse"
      },
...
      "SolutionCreate": {
        "properties": {
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 5,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 20,
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "team_id",
          "title",
          "description"
        ],
        "title": "SolutionCreate"
      },
      "SolutionResponse": {
        "properties": {
          "solution_id": {
            "type": "string",
            "format": "uuid",
            "title": "Solution Id"
          },
          "problem_id": {
            "type": "string",
            "format": "uuid",
            "title": "Problem Id"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "proposed_by": {
            "type": "string",
            "format": "uuid",
            "title": "Proposed By"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/SolutionStatus"
          },
          "industry_review_status": {
            "$ref": "#/components/schemas/IndustryReviewStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "solution_id",
          "problem_id",
          "team_id",
          "proposed_by",
          "title",
          "description",
          "status",
          "industry_review_status",
          "created_at"
        ],
        "title": "SolutionResponse"
      },
      "SolutionStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "SUBMITTED",
          "INDUSTRY_REVIEW",
          "APPROVED",
          "REJECTED"
        ],
        "title": "SolutionStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}